cleanup a bit
