+++ src/server/pkg/cmdutil/cobra.go
@@ -124,77 +124,77 @@ func (r *RepeatedStringArg) Type() string {
 }
 
 func MakeBatchCommand(
-    positionalCount int,
-    cmd *cobra.Command,
-    provision func(*cobra.Command),
-    processFlags func (),
-    run func ([][]string) error,
+	positionalCount int,
+	cmd *cobra.Command,
+	provision func(*cobra.Command),
+	processFlags func (),
+	run func ([][]string) error,
 ) {
-    provision(cmd)
-    cmd.Run = func(cmd *cobra.Command, _ []string) {
-        // Remove non parameter args from the original args so we can reparse
-        // them iteratively with cobra.
-        names := []string{cmd.Name()}
-        names = append(names, cmd.Aliases...)
-        var startArg int
-        for i, x := range os.Args {
-            for _, y := range names {
-                if x == y {
-                    startArg = i + 1
-                    break
-                }
-            }
-        }
-        args := os.Args[startArg:]
+	provision(cmd)
+	cmd.Run = func(cmd *cobra.Command, _ []string) {
+		// Remove non parameter args from the original args so we can reparse
+		// them iteratively with cobra.
+		names := []string{cmd.Name()}
+		names = append(names, cmd.Aliases...)
+		var startArg int
+		for i, x := range os.Args {
+			for _, y := range names {
+				if x == y {
+					startArg = i + 1
+					break
+				}
+			}
+		}
+		args := os.Args[startArg:]
 
-        // Partition by sets of positional args, flags apply to the previous positional args
-        sets := [][]string{{}}
-        index := 0
-        count := 0
+		// Partition by sets of positional args, flags apply to the previous positional args
+		sets := [][]string{{}}
+		index := 0
+		count := 0
 
-        for _, x := range args {
-            if strings.HasPrefix(x, "-") {
-                sets[index] = append(sets[index], x)
-                // TODO: this only works because all flags require a value, otherwise we'll need to check individual flags
-                if !strings.Contains(x, "=") {
-                    count -= 1
-                }
-            } else if count < positionalCount {
-                count += 1
-                sets[index] = append(sets[index], x)
-            } else {
-                sets = append(sets, []string{x})
-                index += 1
-                count = 0
-            }
-        }
+		for _, x := range args {
+			if strings.HasPrefix(x, "-") {
+				sets[index] = append(sets[index], x)
+				// TODO: this only works because all flags require a value, otherwise we'll need to check individual flags
+				if !strings.Contains(x, "=") {
+					count -= 1
+				}
+			} else if count < positionalCount {
+				count += 1
+				sets[index] = append(sets[index], x)
+			} else {
+				sets = append(sets, []string{x})
+				index += 1
+				count = 0
+			}
+		}
 
-        parsedSets := [][]string{}
+		parsedSets := [][]string{}
 
-        // Create an inner command for parsing individual commands
-        innerCommand := &cobra.Command{}
-        provision(innerCommand)
+		// Create an inner command for parsing individual commands
+		innerCommand := &cobra.Command{}
+		provision(innerCommand)
 
-        // Copy over inherited flags so we don't choke on them
-        ancestor := cmd.Parent()
-        for ancestor != nil {
-            innerCommand.Flags().AddFlagSet(ancestor.Flags())
-            ancestor = ancestor.Parent()
-        }
+		// Copy over inherited flags so we don't choke on them
+		ancestor := cmd.Parent()
+		for ancestor != nil {
+			innerCommand.Flags().AddFlagSet(ancestor.Flags())
+			ancestor = ancestor.Parent()
+		}
 
-        // Set a run function that appends to our set of parsed args
-        innerCommand.Run = func (runCmd *cobra.Command, runArgs []string) {
-            parsedSets = append(parsedSets, runArgs)
-        }
+		// Set a run function that appends to our set of parsed args
+		innerCommand.Run = func (runCmd *cobra.Command, runArgs []string) {
+			parsedSets = append(parsedSets, runArgs)
+		}
 
-        for _, argSet := range sets {
-            innerCommand.SetArgs(argSet)
-            innerCommand.Execute()
-            processFlags()
-        }
+		for _, argSet := range sets {
+			innerCommand.SetArgs(argSet)
+			innerCommand.Execute()
+			processFlags()
+		}
 
-        fmt.Printf("flags: %s\n", innerCommand.LocalFlags())
+		fmt.Printf("flags: %s\n", innerCommand.LocalFlags())
 
-        run(parsedSets)
-    }
+		run(parsedSets)
+	}
 }
